+++ packages/vsce/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 
 - Enhancement: Inquire program, Show Library and Inquire Transaction can open respective resources from Resource Inspector. [#449](https://github.com/zowe/cics-for-zowe-client/issues/449)
 - BugFix: Region filtering selects only regions. [#575] (https://github.com/zowe/cics-for-zowe-client/issues/575)
+- Enhancement: Datasets,USS and logs will have hyperlink in RI highlights. [#621](https://github.com/zowe/cics-for-zowe-client/issues/621)
 
 ## `3.20.0`